Refactored service folder, added age to user and added comments
