+++ app/src/main/java/com/example/resdesux2/Server/BoundActivity.java
@@ -1,4 +1,4 @@
-package com.example.resdesux2;
+package com.example.resdesux2.Server;
 
 import android.content.ComponentName;
 import android.content.Context;
@@ -8,12 +8,13 @@ import android.os.IBinder;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.resdesux2.Services.ServerService;
-
 public class BoundActivity extends AppCompatActivity {
     protected ServerService serverService;
     protected boolean isBound;
 
+    /**
+     * This instantiate a connection with the service and handles that it connects.
+     */
     protected final ServiceConnection connection = new ServiceConnection() {
         @Override
         public void onServiceConnected(ComponentName className, IBinder service) {
@@ -29,6 +30,10 @@ public class BoundActivity extends AppCompatActivity {
         }
     };
 
+    /**
+     * Called when activity is started, after onCreate.
+     * This connects and if needed start the Service.
+     */
     @Override
     protected void onStart() {
         super.onStart();
@@ -43,14 +48,25 @@ public class BoundActivity extends AppCompatActivity {
         bindService(intent, connection, Context.BIND_AUTO_CREATE);
     }
 
+    /**
+     * This method is called once the server service is bound and connected to the server,
+     * You can override it and put all your server related calls here.
+     * @param connected This parameter can be neglected as it should always be true.
+     */
     protected void onConnected(boolean connected) {
 
     }
 
+    /**
+     * This method will be called once the activity has been bound to the server service.
+     */
     private void onBound() {
         serverService.setConnectionListener(this::onConnected);
     }
 
+    /**
+     * Called when the activity stopped, it disconnects the service.
+     */
     @Override
     protected void onStop() {
         super.onStop();
